dpdcei: Affine thread for FQDAN setup
The main thread of the application calls the dpdcei setup functions.
The dpdcei setup functions setup the FQDAN (notifications) through the
dpio portals. The thread sending the management command to setup the
FQDAN should always be running on the CPU on which the notification is
setup to be received. This commit correctly affines the main thread so
it is attached to the correct CPU when calling the setup functions

Signed-off-by: Ahmed Mansour <ahmed.mansour@nxp.com>