dce: Enable abort() and discard()
The abort() and discard() functions did not function as expected
because the original implementation relied on the uspc bit in the
abort() case to reset flow context. It was discovered that the DCE
ignores the uspc bit in Recycle=1 recovery frame from suspend. This
was to allow implementations to resend a frame that originally had
the uspc bit set, but was suspended and is now being resent to
continue work

Signed-off-by: Ahmed Mansour <ahmed.mansour@nxp.com>