perf-test: Error out in stateless decomp mismatch
Error out in case the application caller gave it an input file that
does not match the specified chunk size. Previously this would have
resulted in users getting a decompression error, since DCE
decompression will not have read the DEFLATE end block at the end of
the chunk. This results in the error `input starved'. Test team is not
familiar with this mechanism and repeatedly ran into this issue.
This patch makes the test check for this condition and error out early

Signed-off-by: Ahmed Mansour <ahmed.mansour@nxp.com>